# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ship API",
    "version": "1.0.0",
    "description": "Ship is a project and sprint management platform with real-time collaboration.\n\n## Authentication\n\nAll endpoints (except /auth/login and /health) require authentication via:\n- **Session Cookie**: Automatically set after login (15-minute timeout, 12-hour absolute)\n- **Bearer Token**: API tokens from Settings > API Tokens\n\n## Core Concepts\n\n- **Documents**: Everything in Ship is a document (wikis, issues, projects, sprints, etc.)\n- **Document Type**: Each document has a `document_type` that determines its properties\n- **Belongs To**: Documents can be associated via `belongs_to` array (programs, projects, sprints, parent issues)\n\n## WebSocket Collaboration\n\nReal-time editing available at `/collaboration/{docType}:{docId}` using Yjs CRDT protocol.",
    "contact": {
      "name": "Ship Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token authentication. Get your token from Settings > API Tokens."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication. Automatically set after login."
      }
    },
    "schemas": {
...
      "UserReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "User display name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "comment_id": {
            "type": "string",
            "format": "uuid",
            "description": "Thread identifier matching TipTap mark commentId",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "parent_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Parent comment ID for replies, null for root comments",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "content": {
            "type": "string"
          },
          "resolved_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "When the thread was resolved, null if unresolved",
            "example": "2025-01-30T14:30:00.000Z"
          },
          "author": {
            "$ref": "#/components/schemas/UserReference"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 datetime string",
            "example": "2025-01-30T14:30:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 datetime string",
            "example": "2025-01-30T14:30:00.000Z"
          }
        },
        "required": [
          "id",
          "document_id",
          "comment_id",
          "parent_id",
          "content",
          "resolved_at",
          "author",
          "created_at",
          "updated_at"
        ]
      },
...
      "UpdateComment": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "ISO 8601 datetime string",
                "example": "2025-01-30T14:30:00.000Z"
              },
              {
                "nullable": true
              }
            ],
            "description": "Set to ISO datetime to resolve, null to un-resolve"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/comments/{id}": {
      "patch": {
        "operationId": "update_comment",
        "summary": "Update a comment (edit content or resolve/un-resolve)",
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID identifier",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "description": "UUID identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated comment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_comment",
        "summary": "Delete a comment",
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID identifier",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "description": "UUID identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}